# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Background Worker Platform",
    "version": "0.14.0"
  },
  "paths": {
    "/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "description": "Submit a new job.",
        "operationId": "create_job_jobs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_jobs_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max records to return",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "description": "Return job status for polling.",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/file": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job File",
        "description": "Return metadata for the file attached to this job, if any.",
        "operationId": "get_job_file_jobs__job_id__file_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Upload File",
        "description": "Upload a file for later use in a job.",
        "operationId": "upload_file_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/uploads/job": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Create Job With Upload",
        "description": "Upload a file and create an OCR or transcription job.",
        "operationId": "create_job_with_upload_uploads_job_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_job_with_upload_uploads_job_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/uploads/{file_id}": {
      "get": {
        "tags": [
          "uploads"
        ],
        "summary": "Get Upload Metadata",
        "operationId": "get_upload_metadata_uploads__file_id__get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/documents/ingest": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Ingest Document",
        "operationId": "ingest_document_documents_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/chunks": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document Chunks",
        "operationId": "get_document_chunks_documents__document_id__chunks_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rag/query": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Rag Query Async",
        "operationId": "rag_query_async_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rag/query/sync": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Rag Query Sync",
        "description": "Run a RAG query synchronously and return the answer in one response.",
        "operationId": "rag_query_sync_rag_query_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGQuerySyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_create_job_with_upload_uploads_job_post": {
        "properties": {
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "priority": {
            "$ref": "#/components/schemas/JobPriority",
            "default": "normal"
          }
        },
        "type": "object",
        "required": [
          "job_type",
          "file"
        ],
        "title": "Body_create_job_with_upload_uploads_job_post"
      },
      "Body_upload_file_uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "purpose": {
            "$ref": "#/components/schemas/FilePurpose"
          }
        },
        "type": "object",
        "required": [
          "file",
          "purpose"
        ],
        "title": "Body_upload_file_uploads_post"
      },
      "ChunkListResponse": {
        "properties": {
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/ChunkResponse"
            },
            "type": "array",
            "title": "Chunks"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          }
        },
        "type": "object",
        "required": [
          "chunks",
          "total",
          "document_id"
        ],
        "title": "ChunkListResponse"
      },
...
      "DocumentIngestRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "description": "Human-readable name for this document"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "The full text content to ingest"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Where this document came from: 'text', 'upload', 'url'",
            "default": "text"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Optional metadata (author, department, tags, etc.)"
          },
          "chunk_size": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 50.0,
            "title": "Chunk Size",
            "default": 512
          },
          "chunk_overlap": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Chunk Overlap",
            "default": 50
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "DocumentIngestRequest",
        "description": "Request body for document ingestion."
      },
      "DocumentIngestResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentResponse"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Document submitted for ingestion"
          }
        },
        "type": "object",
        "required": [
          "document",
          "job_id"
        ],
        "title": "DocumentIngestResponse",
        "description": "Document record plus job ID for ingestion status polling."
      },
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "metadata_": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "chunk_size": {
            "type": "integer",
            "title": "Chunk Size"
          },
          "chunk_overlap": {
            "type": "integer",
            "title": "Chunk Overlap"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "source",
          "status",
          "chunk_size",
          "chunk_overlap",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse"
      },
...
      "FileUploadResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "purpose": {
            "$ref": "#/components/schemas/FilePurpose"
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated",
            "description": "True if this upload matched an existing file by hash"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "original_filename",
          "file_type",
          "file_size",
          "content_hash",
          "purpose",
          "deduplicated",
          "created_at"
        ],
        "title": "FileUploadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreate": {
        "properties": {
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input",
            "description": "Job-specific input payload"
          },
          "priority": {
            "$ref": "#/components/schemas/JobPriority",
            "default": "normal"
          }
        },
        "type": "object",
        "required": [
          "job_type",
          "input"
        ],
        "title": "JobCreate"
      },
      "JobFileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "purpose": {
            "$ref": "#/components/schemas/FilePurpose"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "original_filename",
          "file_type",
          "file_size",
          "content_hash",
          "purpose",
          "created_at"
        ],
        "title": "JobFileResponse"
      },
      "JobListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "type": "array",
            "title": "Jobs"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "jobs",
          "total"
        ],
        "title": "JobListResponse"
      },
...
      "JobPriority": {
        "type": "string",
        "enum": [
          "high",
          "normal",
          "low"
        ],
        "title": "JobPriority"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/JobPriority"
          },
          "input_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input Payload"
          },
          "result_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Payload"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_type",
          "status",
          "priority",
          "input_payload",
          "result_payload",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed"
        ],
        "title": "JobStatus"
      },
      "JobType": {
        "type": "string",
        "enum": [
          "summarization",
          "ocr",
          "embeddings",
          "transcription",
          "recommendations",
          "ingestion",
          "rag_query",
          "query_expand",
          "rerank",
          "rag_retrieve",
          "rag_generate"
        ],
        "title": "JobType"
      },
...
      "RAGQueryResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "RAG query submitted"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "inline"
          },
          "step_job_ids": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Job Ids"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "RAGQueryResponse",
        "description": "Immediate ack with job ID for status polling."
      },
...
    }
  }
}